Use gosu and map eclipse user
Map eclipse user to the user owning the workspace. Most of the time the
end user will mount it's projects on the workspace directory. Since
the container writes to this volume as the eclipse user it can result in
files being owned by a strange uid.

To ensure all files in the workspace are still owned by the workspace
owner I'll map the uid/gid of the eclipse user in the container to the
owner of the possible mounted workspace volume.